Use an entire register for inputs/output
Tests GPU code now takes input 1 from t2 and input 2 from t3, and outputs to t4.
The entire 4x4-byte vectors are read from the input and written to the
output.

This makes it possible to test the 4-wide behavior of instructions, not
just scalar operations. It also makes the test code easier to read.